change str_replace not working properly
I've trying retrieve multiple CallerID Numbers as filter without sucess,
Checked it using regexp "_200[67]" as CallerID Number without success. 
The $cnum variable showed as:
"AND (cnum RLIKE '^200[67]$' OR src AND cnum RLIKE '^200[67]$' )"
Note "AND cnum" still there.

Then, i've use "2006,2007" but only 2006 showed in results.
The $cnum variable showed as:
"AND (cnum LIKE '2006%' OR cnum LIKE '2007%' OR src LIKE '2006%' OR cnum LIKE '2007%')"
Take a look at last cnum 2007. This must be src but still as cnum.

With my proposition, this will not occur since we are getting string position.
Isn't clear the root cause of that but I think it's due to string manipulation before str_replace.